Modularize queries, mutations, and types
